agentic-pharma-ai/requirements-api.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-core==0.1.0
langchain-openai==0.0.5
langgraph==0.0.1
python-dotenv==1.0.0
requests==2.31.0
